requirements.txt
langchain-openai
langchain_core
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
PyMuPDF
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
fastapi
langserve